Refractor: Extract IndexFilter as a seperate component
